moved button function to challenges
