--- Components/ShareScreen.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import { Text } from 'react-native'
-
-
-export const ShareScreen = () => {
-    return (
-        <Text>
-            share screen text
-        </Text>
-
-    )
-}
