+++ Components/StartScreenStyling.js
@@ -1,14 +1,21 @@
 import styled from 'styled-components/native';
 
+export const Screen = styled.View`
+  background-color: papayawhip;
+`
 
 export const Title = styled.Text`
   font-size: 24px;
-  color: palevioletred;
+  color: palevioletred; 
 `
 
-
 export const IntroText = styled.Text`
-max-width: 500px;
-text-align: center;
-padding-top:30px;
+  max-width: 500px;
+  text-align: center;
+  padding-top:30px;
+`
+
+export const ResultText = styled.Text`
+  padding:20px;
+  font-size:20px;
 `
